+++ cypress/support/helper.js
@@ -26,7 +26,7 @@ export function visitComponentUrl(
   env = "story"
 ) {
   cy.visit(prepareUrl(component, suffix, iFrameOnly, prefix, env));
-  if (!iFrameOnly) knobsTab().click();
+  if (!iFrameOnly) knobsTab().click({ force: true });
 }
 
 // eslint-disable-next-line max-params
@@ -53,7 +53,7 @@ export function visitComponentUrlWithParameters(
 }
 
 export function clickActionsTab(iFrameOnly = false) {
-  if (!iFrameOnly) actionsTab().click();
+  if (!iFrameOnly) actionsTab().click({ force: true });
 }
 
 export function clickClear() {