+++ src/components/portrait/portrait-initials.component.js
@@ -1,5 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
+import styledSystemPropTypes from "@styled-system/prop-types";
 import { withTheme } from "styled-components";
 import sizeParams from "./portrait-size.config";
 import BaseTheme from "../../style/themes/base";
@@ -10,6 +11,12 @@ import {
   StyledPortraitInitialsImg,
   getColorsForInitials,
 } from "./portrait.style";
+
+import { filterStyledSystemMarginProps } from "../../style/utils";
+
+const marginPropTypes = filterStyledSystemMarginProps(
+  styledSystemPropTypes.space
+);
 
 class PortraitInitials extends React.Component {
   /** Cache of the initials graphic. */
@@ -111,6 +118,7 @@ class PortraitInitials extends React.Component {
 }
 
 PortraitInitials.propTypes = {
+  ...marginPropTypes,
   /** The theme to use. */
   theme: PropTypes.object,
   /** The user's initials to render. */